refactor: moved all server packets in new structure in mojang/minecraft/net/packets/*
