+++ mojang/minecraft/net/packets/status/server/ping_request.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from mojang.minecraft.net.packet import Packet
+
+
+@dataclass
+class PingRequest(Packet):
+    packet_id = 0x01
